Removed from move to damage event in cause of void
